requirements.txt
attrs==25.4.0
cattrs==25.3.0
certifi==2025.11.12
charset-normalizer==3.4.4
contourpy==1.3.3
cycler==0.12.1
fastf1==3.6.1
fonttools==4.60.1
idna==3.11
influxdb-client==1.49.0
kafka-python==2.2.15
kiwisolver==1.4.9
matplotlib==3.10.7
numpy==2.3.4
packaging==25.0
pandas==2.3.3
pillow==12.0.0
platformdirs==4.5.0
pyparsing==3.2.5
python-dateutil==2.9.0.post0
pytz==2025.2
RapidFuzz==3.14.3
reactivex==4.1.0
requests==2.32.5
requests-cache==1.2.1
scipy==1.16.3
setuptools==80.9.0
six==1.17.0
timple==0.1.8
typing_extensions==4.15.0
tzdata==2025.2
url-normalize==2.2.1
urllib3==2.5.0
websockets==13.1
